kumo-screenshot-worker: move diffing back into CI
